removed size_t at all
git-svn-id: https://btanks.svn.sourceforge.net/svnroot/btanks/trunk@837 5c0bf991-d514-0410-a65f-ce4832e571f3

+++ mrt/serializator.cpp
@@ -109,7 +109,7 @@ void Serializator::add(const Chunk &c) {
 void Serializator::add(const float f) {
 	//LOG_DEBUG(("added float %f", f));
 	char buf[256];
-	size_t len = snprintf(buf, sizeof(buf) -1, "%g", f);
+	unsigned int len = snprintf(buf, sizeof(buf) -1, "%g", f);
 	add(std::string(buf, len));
 }
 
@@ -152,7 +152,7 @@ void Serializator::get(float &f) const {
 }
 
 void Serializator::get(std::string &str)  const {
-	size_t size;
+	unsigned int size;
 	get(size);
 
 	ASSERT_POS(size);
